requirements.txt
apscheduler>=3.9.1
python-telegram-bot>=20.4
python-dotenv>=1.1.0
pyodbc>=4.0.40
